Chances of system crash on boot ~70% -> ~40%

+++ include/kernel/bitmap.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <typedef.h>
 typedef struct bitmap {
-    uint32_t* bytes;
+    uint8_t* bytes;
     uint64_t size;
     bool operator[](uint64_t index);
     void set(uint64_t index, bool value);
